LayoutTests/fast/forms/datetimelocal/datetimelocal-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for datetime-local input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, {min or max}, [stepCount]).
Normal cases
PASS stepUp("2010-02-10T20:13", null, null) is "2010-02-10T20:14"
PASS stepDown("2010-02-10T20:13", null, null) is "2010-02-10T20:12"
PASS stepUp("2010-02-10T20:13", null, null, 10) is "2010-02-10T20:23"
PASS stepDown("2010-02-10T20:13", null, null, 11) is "2010-02-10T20:02"
PASS stepUp("1970-01-01T20:13", "4", null, 2) is "1970-01-01T20:13:08"
PASS stepDown("1970-01-01T20:13", "4", null, 3) is "1970-01-01T20:12:48"
Step=any
PASS stepUp("2010-02-10T20:13", "any", null) is "2010-02-10T20:14"
PASS stepDown("2010-02-10T20:13", "any", null) is "2010-02-10T20:12"
Overflow/underflow
PASS stepUp("2010-02-10T20:13", "3.40282346e+38", null) is "275760-09-13T00:00:00.000"
PASS stepDown("2010-02-10T20:13", "3.40282346e+38", null) is "1970-01-01T00:00:00.000"
PASS stepUp("2010-02-10T20:13", "1", "2010-02-10T20:13") is "2010-02-10T20:13"
PASS stepDown("2010-02-10T20:13", "1", "2010-02-10T20:13") is "2010-02-10T20:13"
stepDown()/stepUp() for stepMismatch values
PASS stepDown("2010-02-10T20:13", "3", "2010-02-10T20:12:56") is "2010-02-10T20:12:59"
PASS stepUp("1970-01-01T00:13", "7", "") is "1970-01-01T00:13:04"

PASS successfullyParsed is true

TEST COMPLETE
